% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentMetaMean.R
\name{findHDI}
\alias{findHDI}
\title{Find the highest density interval(s)}
\usage{
findHDI(l, p = 0.95, stepsize = 0.01, density_resolution = 0.001,
  drop = T)
}
\arguments{
\item{l}{a result list returned by \code{studentMetaMean}}

\item{p}{a single percentage or a vector of percentages for which to return the HDI}

\item{stepsize}{size of steps along the x axis}

\item{density_resolution}{cutoff for sampling the distribution, given as the fraction of
the density at the mean}

\item{drop}{return a vector instead of a matrix for a single target percentage}
}
\value{
Either a vector (low, high) or a matrix with each row containing a (low, high) pair
}
\description{
Find the highest density interval(s)
}

