% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentMetaMean.R
\name{plotMetaMean}
\alias{plotMetaMean}
\title{Plot both the individual distributions and the combined distribution.}
\usage{
plotMetaMean(l, p = 0.95, stepsize = NULL, xmin = NULL, xmax = NULL)
}
\description{
Plot both the individual distributions and the combined distribution.
}

