% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentMetaMean.R
\name{pMetaMean}
\alias{pMetaMean}
\title{Compute a value in the cumulative distribution function}
\usage{
pMetaMean(l, p, return_pvalue = F)
}
\description{
Compute a value in the cumulative distribution function
}

