% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentMetaMean.R
\name{studentMetaMean}
\alias{studentMetaMean}
\title{Compute a weighted average of the specified distributions}
\usage{
studentMetaMean(design = NULL, means = NULL, sigmas = NULL, nus = NULL,
  min_sigma = 0.01)
}
\arguments{
\item{design}{A matrix with the named columns \code{mean}, \code{sigma}, and \code{nu}. Can be omitted,
if the following three parameters are supplied.}

\item{means}{Vector of means}

\item{sigmas}{Vector of sigmas}

\item{nus}{Vector of nus}

\item{min_sigma}{Minimum sigma of resulting distribution when all means are very close together}
}
\value{
A list with the following values:
 \itemize{
   \item{\code{xmin}}{ effective lower end of probality distribution}
   \item{\code{xmax}}{ effective upper end of probality distribution}
   \item{\code{mean}}{ weighted mean}
   \item{\code{pMerged}}{ combined probability distribution}
   \item{\code{likelihood_functions}}{ the individual likelihood function for the supplied distributions}
 }
}
\description{
Compute a weighted average of the specified distributions
}

